dev-requirements.txt
-r requirements.txt
python-dotenv==1.0.1
django-extensions